fix: amp

+++ ds/loader.py
@@ -87,6 +87,6 @@ class DataLoader:
 
         def jax_iterator():
             for item in self.dataset:
-                yield make_batch(jnp.asarray(item, dtype=jnp.int32))
+                yield make_batch(jnp.asarray(item))
 
         return jax_iterator()
